Linear Regression(Prodigy_Internship)
Linear regression is a statistical technique used for modeling the relationship between a dependent variable and one or more independent variables. In the context of predicting house prices, linear regression can be employed to estimate the relationship between various features of a house (such as square footage, number of bedrooms, and number of bathrooms) and its price.